+++ src/clients/carbon-client/carbon-client-state.enum.ts
@@ -1,0 +1,5 @@
+export enum CarbonClientState {
+    CONNECTED = 'CONNECTED',
+    CONNECTING = 'CONNECTING',
+    DISCONNECTED = 'DISCONNECTED'
+}